pyproject.toml
[project]
name = "emotionful-bot"
version = "0.1.0"
description = "An advanced Telegram stickers/emojis manager tool wrapped in Telegram Bot."
authors = [
    {name = "frost-official"},
]
maintainers = [
    {name = "frost-official"},
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "aiogram[i18n] (>=3.17.0,<4.0.0)",
    "tortoise-orm (>=0.23.0,<0.24.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "aiosqlite (>=0.20.0,<0.21.0)"
]
classifiers = [
  "Development Status :: 3 - Alpha",

  "Private :: Do Not Upload",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]

[project.urls]
Telegram = "https://EmotionfulBot.t.me"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"